refactor: move combat UI helpers
